+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   `delete_at`, `resize`, `reverse`, `rotate, and `sort`.
 - New methods for `RString`: `enc_new`, `len`, `length`, and `is_empty`.
 - `RHash` gains the methods `delete` and `clear`.
+- `require` function
 
 ### Changed
 - `RArray::cat`, `RArray::from_slice`, and `gc::mark_slice` will accept a